IMPALA-7121: Clean up partitionIds_ from HdfsTable
The purpose of introducing partitionIds_ member to HdfsTable was to be
able to return the IDs of all the current partitions in constant time.
Apparently, partitionMap_ also contains these IDs as the key of the
map and this is accessible via keySet() also in constant time. It
seems reasonable then to remove partitionIds_ and use
partitionMap_.keySet() in getPartitionIds() to save some memory.

          ==C5_FEATURE_IMPALA_METADATA==

Change-Id: I8b5a480e570aeae565fafd4f3e2b279e7a98c7da
Reviewed-on: http://gerrit.cloudera.org:8080/10654
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit 2cc2f828799cc67de2e8ee0c74b037f171820bde)