IMPALA-7128 (part 1) Refactor interfaces for Db, View, Table, Partition
This refactors out interfaces in the frontend for the interaction
between the analysis/planning code and the classes that implement
these catalog objects.

This takes care of the most commonly used objects but defers some others
(e.g. functions, cache pools, data sources, etc) to follow-on patches.

There are a few spots remaining in the frontend that still downcast to
implementation classes, particularly where the frontend actually makes
modifications to catalog objects in-place. I left TODOs in those spots
and will come back later as necessary.

Change-Id: Id55f7d2e94d81e66ce720acb6315f15a89621b31
Reviewed-on: http://gerrit.cloudera.org:8080/10611
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>