IMPALA-7135. Skeleton implementation of LocalCatalog
This adds some of the high level classes for implementing the local
catalog:

- LocalCatalog is the top level implementation. The plan is to
  instantiate this once per query, so that no thread safety is required.

- It loads metadata from a MetaProvider interface. The current
  implementation fetches directly from HMS and provides no caching. A
  future subtask will add a CachingMetaProvider implementation.
  Separating out caching will make it easier to experiment with
  different policies or storage mechanisms.

- It instantiates LocalDb and LocalTable objects to implement FeDb and
  FeTable. These are mostly stubbed out except for the most basic
  functionality. Functionality will be filled in incrementally in
  further patches.

Since it's not yet possible to hook this up to most of the existing
tests, a very simple new unit test is included to cover the bits of
functionality that are not stubbed out. I didn't concentrate on too much
test coverage here, since once we've implemented more functionality we
can switch over all of the existing tests to get coverage of the new
implementation.

Change-Id: Iab653371188b21c72f50ee1ec4e94950aa6fb9ee
Reviewed-on: http://gerrit.cloudera.org:8080/10627
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>