IMPALA-7137. Support configuring Frontend to use LocalCatalog
This adds a new flag -use_local_catalog which is passed through to the
frontend and causes it to use LocalCatalog instead of ImpaladCatalog.
Additionally, when this flag is configured, the impalad does not
subscribe to catalog topic updates from the statestore.

The patch is slightly more complex than simply picking which class to
instantiate, because the lifecycle is designed a bit differently between
the two implementations:

- LocalCatalog is instantiated once per query/request.

- ImpaladCatalog is instantiated once and stateful across queries,
  except when a full catalog update is received. This maintains the
  current behavior for this implementation.

In order to abstract this difference in lifecycle from the frontend, I
introduced a new FeCatalogManager class with different implementations
for the two lifecycles. I also had to add a simple test implementation
since some tests rely on directly injecting a Catalog implementation
into the Frontend.

This patch also includes a few small changes to the local catalog
implementation objects to enable the impalad to start and accept
connections. With this patch, I was able to manually test as follows:

I started just the statestore and the impalad in the new mode:

- ./bin/start-statestored.sh
- ./bin/start-impalad.sh --use_local_catalog

I connected with impala-shell as usual and was able to run the most
simple queries:

- SHOW DATABASES;
- USE functional;
- SHOW TABLES;

All other functionality results in error messages due to the various
TODOs in the current skeleton implementation.

Change-Id: I8c9665bd031d23608740b23eef301970af9aa764
Reviewed-on: http://gerrit.cloudera.org:8080/10629
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Vuk Ercegovac <vercegovac@cloudera.com>