IMPALA-7140 (part 2). Create skeleton for LocalFsTable
This adds a skeleton implementation of FeFsTable for the LocalCatalog.
Most of the implementation in this patch is stubbed out -- only the bare
bones necessary to instantiate the subclass when a table is loaded and
the simplest table-level methods.

          ==C5_FEATURE_IMPALA_METADATA==

Change-Id: Id2b184104d92e128250df5a08ac7ffb3dde011a8
Reviewed-on: http://gerrit.cloudera.org:8080/10712
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Todd Lipcon <todd@apache.org>
(cherry picked from commit 94e713081b562e5870e2c339892a807dd1a84228)