IMPALA-7140 (part 4): support creating descriptors for FS tables
This adds the relevant methods to convert LocalFsTable and
LocalFsPartition to thrift descriptors for consumption by the backend.

Unfortunately we cannot yet enable the planner tests, since they are
checking file counts and sizes as part of the explain output, and we
haven't yet implemented file info fetching in the LocalCatalog.

However, I was able to manually test this change by starting an impalad
with --use_local_catalog, connecting to it from the shell, and running
various EXPLAIN SELECT queries against tpch and functional tables. The
explain output is more or less as expected with the exception of missing
file info.

          ==C5_FEATURE_IMPALA_METADATA==

Change-Id: I4550612eb6d1e3a324f49a9c4d24b048e45d3738
Reviewed-on: http://gerrit.cloudera.org:8080/10735
Reviewed-by: Vuk Ercegovac <vercegovac@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit bd52a751bcde6a62aed4c0f9536c64b4536c9a87)