IMPALA-7151: Rework ephemeral port assignment for be tests
Instead of using FindUnusedEphemeralPort() to choose
many ephemeral ports for in-process servers, we use
Thrift's builtin wildcard ports, where it picks an
ephemeral port when the configured port is 0. This
should be more robust because there's no window where
something else can steal our port. Only Thrift 0.9.2+
support the getPort() method required for this.

This does not necessarily make all ImpalaServer
functionality work with ports configured to 0,
just enough to get expr-test and session-expiry-test
to work. E.g. various strings and names get port 0 in
them and query execution may not work end-to-end.

Some tests still use FindUnusedEphemeralPort() to
pick a single port at a time. We haven't seen that usage pattern
to be flaky with any frequency and it appears to be more work
to fix those, so I left them for now.

Testing:
Ran core tests.

Cherry-picks: not for 2.x

Change-Id: I2fcc9243099f8249380ac5f01ac5eb67cba24cf5
Reviewed-on: http://gerrit.cloudera.org:8080/10740
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>