IMPALA-7150: fixes jni frame scope
A previous change (https://gerrit.cloudera.org/#/c/9968/)
reduced the time that a lib-cache entry was held when initializing
a udf. It did so by introducing a new scope. However, that scope
captured a jni frame, which resulted in an earlier frame pop
than before. This change decouples the frame scope from the lib-cache
scope so that the frame is popped at the end of the method (as before)
instread of with the lib-cache scope.

Testing:
- core, asan tests pass

Change-Id: I38d482dde6ed7f8857ae565bf62a8e923ffb0ae9
Reviewed-on: http://gerrit.cloudera.org:8080/10746
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/35595
Tested-by: Tim Armstrong <tarmstrong@cloudera.com>
Reviewed-by: Philip Zeyliger <philip@cloudera.com>