IMPALA-7175: deflake check for failed impalad
test_native_functions_race checks for an impalad crash
by testing whether the number of impalads at the start and
end of the test is the same. A recent run was flaky since
the number of impalads at the start was incorrectly found
to be 2. This fix tries to make the test most robust by
determining the number of impalads based on how many
can evaluate a trivial test query. For these tests, its
assumed that the number of coordinators is the same as
the number of impalads in the cluster.

Change-Id: I97c6b398e43c6abb1df2b1783c26159137f14fa4
Reviewed-on: http://gerrit.cloudera.org:8080/10745
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit 03dd65d793524dd4d8bad90efd9d1de4f6a4cd82)
Reviewed-on: https://gerrit.sjc.cloudera.com/37035
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>