IMPALA-7277. Support INSERT and LOAD DATA statements in LocalCatalog
This adds support for INSERT and LOAD DATA statements when LocalCatalog
is enabled by fixing the following items:

* Remove some downcasts to HdfsTable in various Stmt classes and replace
  them with casts to FeFsTable.
* Stub out the "write access" checks to fake always being writable. Left
  a TODO to properly implement these.
* Implemented various 'getPartition()' calls which take a user-provided
  partition specification, used by INSERT and LOAD statements that
  specify an explicit target partition.
* Fixed the "prototype partition" created by LocalFsTable to not include
  a location field. This fixed insertion into dynamic partitions.

Additionally fixed a couple other issues which were exercised by the
e2e test coverage for load/insert:

* The LocalCatalog getDb() and getTable() calls were previously assuming
  that all callers passed pre-canonicalized table names, but that wasn't
  the case. This adds the necessary toLower() calls so that statements
  referencing capitalized table names work.

With this patch, most of the associated e2e tests pass, with the
exception of those that check permissions levels of the target
directories. Those calls are still stubbed out.

Overall, running 'run-tests.py -k "not hbase and not avro"' results in
about a 90% pass rate after this patch:

=====================================================
186 failed, 1657 passed, 56 skipped, 33 xfailed, 1 xpassed in 512.75 seconds
=====================================================

Remaining test failures seem mostly unrelated to the above changes and
will be addressed in continuing patches.

Change-Id: I4ae47a0b58022ed77abe51d2596c2b1d0111fae3
Reviewed-on: http://gerrit.cloudera.org:8080/10914
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Todd Lipcon <todd@apache.org>