IMPALA-7275: Create table authorization error should not show table name
The authorization correctly checks database level authorization but
the exception includes the table name.  This fix adds a check for
create and modifies the error string to remove the table name.  Having
an error with CREATE on a table is not correct since to create a table
a user needs CREATE on the parent database.

Example:
SQL: CREATE TABLE db.tbl(col1 INT)
Before:
User 'user1' does not have privileges to execute 'CREATE' on: db.tbl
After:
User 'user1' does not have privileges to execute 'CREATE' on: db

Testsing: Ran all FE tests

Change-Id: I6711a744541abd5ff26574974ba8517b6e51c453
Reviewed-on: http://gerrit.cloudera.org:8080/10919
Reviewed-by: Fredy Wijaya <fwijaya@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>