IMPALA-6810: runtime_row_filters.test: omit pool name in pattern
Some downstream tests run this with a fair-scheduler.xml set that, while
not changing admission control behavior, does change the name of the
pool. Omit the pool name to permit that downstream test to succeed.

Testing:
- local with change in minicluster
- downstream in environment as well

Conflicts:
  testdata/workloads/functional-query/queries/QueryTest/runtime_row_filters.test

Conflicts exist because "IMPALA-6035: Add query options to limit thread
reservation", which altered the exception pattern, isn't ported to
6.0.x. I chose to resolve the conflicts by just altering the message in
the same "spirit" as upstream IMPALA-6810: trim the front of the
message.

Testing:
- https://master-02.jenkins.cloudera.com/view/Impala/view/Private/job/impala-private-parameterized/2513/
- https://master-02.jenkins.cloudera.com/view/Impala/view/Private/job/impala-private-cdh-cluster-tests/19/

==C5_APPROVED_BUGFIX==

Change-Id: I3fe6beb169dc6bfefabde9dc7a4632c1a5e63fa7
Reviewed-on: https://gerrit.sjc.cloudera.com/37301
Reviewed-by: Zoram Thanga <zoram@cloudera.com>
Tested-by: Nithya Janarthanan <njanarthanan@cloudera.com>
Reviewed-on: https://gerrit.sjc.cloudera.com/39169
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>