IMPALA-7315: fix test_update_with_clear_entries_flag race
We need to wait for the subscriber to process the second update in order
to guarantee that the first update for that subscriber has been applied.
Otherwise there is a race window where the second subscriber may see the
older version of the statestore topic.

==C5_APPROVED_BUGFIX==

Change-Id: I2be2b61b6deb0228fbc5a242e43076beb8871454
Reviewed-on: http://gerrit.cloudera.org:8080/10986
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit 70e2d57fc06499d3eb7c278c28daf51f782d93b3)
Reviewed-on: https://gerrit.sjc.cloudera.com/36247
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>