IMPALA-7386. Replace CatalogObjectVersionQueue with a multiset
The implementation of CatalogObjectVersionQueue was based on a priority
queue, which is not a good choice of data structure for the use case.
This class exposes the following operations, with corresponding runtimes
for the heap:

- addVersion: O(lg n)
- removeVersion: O(n)
- getMinVersion: O(1)

Given that every update of a catalog object requires removing the old
version, the O(n) runtime could get quite expensive, particularly for
operations like INVALIDATE METADATA which end up removing the old
version of all of the objects in the catalog -- thus becoming
accidentally quadratic.

The new patch switches to a TreeMultiset coupled with a simple cache of
the min element, with runtimes:

- addVersion: O(lg n)
- removeVersion: O(lg n)
- getMinVersion: O(1)

The downside of this patch is increased memory overhead: we are going
from a PriorityQueue which has 8 bytes overhead per element to a
TreeMultiset which has 48 bytes overhead per element. In a catalog with
millions of tables this won't be insignificant; however, in such a
catalog the performance savings are probably critical. According to
a quick JMH benchmark, removal of an element from a PriorityQueue with
1M elements takes about 3.5ms, so an operation which invalidates all
million tables would take about 3500 seconds without this patch.
Meanwhile, the same benchmark using a TreeSet takes about 28ns per
removal, so the whole invalidation would take about 28ms.

This patch also makes the data structure synchronized, just for
simplicity's sake, since uncontended locks are very cheap in Java.

  ==C5_FEATURE_IMPALA_METADATA==

Change-Id: I1b6c58a1acef9b02fcc26a04d048ee9cc47dc0ef
Reviewed-on: http://gerrit.cloudera.org:8080/11109
Tested-by: Todd Lipcon <todd@apache.org>
Reviewed-by: Todd Lipcon <todd@apache.org>
(cherry picked from commit b9d44aff3ee5850dc7f08ba61091c959f08bc474)