IMPALA-7388. Fix issues in and optimize various Status macros
Prior to this patch, the JNI THROW_IF macros used a variable 'status' in
local scope to hold the result of the statement whose status is to be
checked. This could cause problems if there is another variable 'status'
in the external scope, as the inner scope would shadow the outer. For
example:

  Status status = foo();
  THROW_IF_ERROR(status, ...);

... would actually end up resulting in a macro expansion like:

  do {
    Status status = status;
    ...
  } while (false);

Suprisingly, such code is legal C++ and ends up calling the Status copy
constructor with an uninitialized version of itself as the argument. gcc
doesn't seem to emit a warning for this construct, though clang does
emit "variable 'status' is uninitialized when used within its own
initialization".

If such code is emitted, we'll get undefined behavior and most likely
not end up throwing an error even if status is not OK.

The fix is simple: use a less-likely-to-conflict variable name _status
in the macro. This also fixes other macros that previously used
__status__ to use _status to align with the fact that __*__ style
keywords are reserved by the C++ standard.

Additionally, this patch changes some other similar macros to use const
references instead of creating new variables. On the surface, it might
look wrong to assign a const reference to a temporary such as the result
of a function call. However, C++ explicitly allows this and ensures that
the lifetime of the temporary is extended to the lifetime of the
captured reference, making it safe[1]. This acts as a small optimization
since the RETURN_IF_ERROR macros won't call their copy-constructor an
extra time.

[1] https://herbsutter.com/2008/01/01/gotw-88-a-candidate-for-the-most-important-const/

Change-Id: I3d62c99bbd83cc290f5992077258af156eaa6330
Reviewed-on: http://gerrit.cloudera.org:8080/11153
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Todd Lipcon <todd@apache.org>
(cherry picked from commit f572bc8cd70b325c5a9114849466da9f277b37cb)
Reviewed-on: https://gerrit.sjc.cloudera.com/37051
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>