IMPALA-7394: Don't print stack trace in ExpireSessions()
When sessions expire, ExpireSessions() may create excessive amount
of log spew due to printing of the stack traces esp when there are
a massive amount of sessions expiring at the same time. This stack
crawling also unnecessarily extends the critical section under the
session_state_map_lock_, increasing the contention for that lock.
This change converts the use of Status() to Status::Expected() to
remove the unnecessary stack crawl.

Change-Id: I29cb143af96e2eef45479a03365d8f49a2ee2dfa
Reviewed-on: http://gerrit.cloudera.org:8080/11117
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>