IMPALA-7335/IMPALA-7418: Add logs to HdfsScanNode to debug the issues
IMPALA-7335 and IMPALA-7418 are failing some builds on jenkins.
However, there is no deterministic method to reproduce them
locally and hence it is difficult to figure out the cause of the
failure. From the existing logs, it appears that the status
generated by HdfsScanNode::ProcessSplit() is lost. This log
would help determine the condition when the failures occur.

Change-Id: I68698c90031edc6ee8c31e9ce3d52dade9d8f6f1
Reviewed-on: http://gerrit.cloudera.org:8080/11174
Reviewed-by: Bikramjeet Vig <bikramjeet.vig@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>