IMPALA-7407, IMPALA-7067: Fix test_cancellation failure on KeyboardInterrupt
test_cancellation runs a shell process, executes a query, sleeps,
sends a sigint to the process, and then checks that the query is
cancelled. If the sigint is sent prior to the shell installing its
signal handler, the test can fail with a KeyboardInterrupt.

This patch removes the reliance on the sleep being long enough by
actually reading the output of the shell and only cancelling the
query once the shell shows that it has started running.

Testing:
- Ran test_cancellation in a loop.

Also includes the below commit since it was the easiest way to resolve
the merge conflict:

--------------------------------------------------------------------
IMPALA-7067: deflake test_cancellation
Tweak the query so that it still runs for a long time but
can cancel the fragment quicker instead of being
stuck in a long sleep() call.
--------------------------------------------------------------------

Change-Id: I65302ffb838d5185f77853bc2e53296f3a701d93
Reviewed-on: http://gerrit.cloudera.org:8080/11255
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Thomas Marshall <thomasmarshall@cmu.edu>
Reviewed-on: https://gerrit.sjc.cloudera.com/c/Impala/+/45591
Tested-by: Tim Armstrong <tarmstrong@cloudera.com>
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>