IMPALA-7442: reduce mem requirement of semi-joins-exhaustive
The test started running into IMPALA-7446, maybe because of
a timing change. This appears to have always been possible.

The fix is to reduce the memory requirement of the test.
IMPALA-2256 is no longer really possible because the
BufferedTupleStream code was simplified to avoid the 32-bit
row index limitation, so we're not losing important coverage
on the current code with this change.

Testing:
Ran test in a loop to confirm it did not OOM.

Change-Id: I9d9480cad6bf8222abe990e7046498a0531e2849
Reviewed-on: http://gerrit.cloudera.org:8080/11223
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit 6bb9af02b87a71cc65c21b5916f8214441917601)
Reviewed-on: https://gerrit.sjc.cloudera.com/37053
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>