IMPALA-7418: Return first non-ok status from HdfsScanNode::GetNext()
Previously, HdfsScanNode::GetNext passed the status returned by
IssueInitialScanRanges() without inspecting the
HdfsScanNodeBase::status_. This resulted in the error status being
lost in case a scanner thread hit an error and cancelled the scan.
This change ensures that GetNext() returns the first non-ok status
set in HdfsScanNode.

Testing: Added sleeps to the IssueInitialRanges() to cause
deterministic failures of test_udf_errors and then applied this
patch to it. It passes all the tests despite the sleeps.

Change-Id: I4569cc7b0843a29c617a094e590c31f7c648ff45
Reviewed-on: http://gerrit.cloudera.org:8080/11296
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>