IMPALA-7335: Fix a race in HdfsScanNode
This change fixes the race between the done_ flag and the status_
variable in the HdfsScanNode. Previously, a scanner thread would
mark its scan range as complete even if it ran into an error.
Another scanner thread could notice that all scan ranges have
completed and set the done_ flag before the status_ variable is
updated with the non-ok status. This caused the main thread to
return an ok status despite the scanner thread running into an
error. This change ensures  that when a scanner thread runs into
an error, it updates the status_ before marking its scan range as
complete.

Testing: Ran core tests.
Since there is no deterministic method to reproduce the
race, this change does not include any regression tests.

Additionally, this change also fixes IMPALA-7430 by removing the
logs added to debug this race.

Change-Id: Id470818846a5c69ad28a6ff695069702982aa793
Reviewed-on: http://gerrit.cloudera.org:8080/11337
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>