IMPALA-7502: ALTER TABLE RENAME should require ALL on the old table
Prior to this patch, ALTER TABLE/VIEW RENAME required ALTER on the old
table. This may pose a potential security risk, such as having ALTER on
a table and ALL on a particular database allows a user to move the table
to a database with ALL, which will automatically grant that user with
ALL privilege on that table due to the privilege inherited from the
database. This patch fixes the issue by requring ALL on the old table.
What this means is moving a table to a database with ALL privilege will
not elevate the privilege since ALL is now required for a table to be
renamed.

Testing:
- Ran all FE tests

==C5_FEATURE_FGP==

Change-Id: I47a417a77df3f3030cf5f54fd2280b5e5e1fb77a
Reviewed-on: http://gerrit.cloudera.org:8080/11344
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-on: https://gerrit.sjc.cloudera.com/36732
Reviewed-by: Vuk Ercegovac <vercegovac@cloudera.com>
Tested-by: Fredy Wijaya <fwijaya@cloudera.com>