CDH-72381: Set -XX:MaxPermSize if not set by the users.
Due to increased class footprint in CDH5.16, we noticed that the tests
are running into perm-gen OOMs and such a scenario is possible in the
real world clusters too.

This commit appends -XX:MaxPermSize=128m to JAVA_TOOL_OPTIONS if it is
not already set by the users. This flag is only applicable to Java
versions <= 1.7. We do not set if if Java version >=1.8

Testing: Manual. With Java 1.7 and 1.8.

Change-Id: I0e1b98985932249a8251b874d1c64de4b485fffc
Reviewed-on: https://gerrit.sjc.cloudera.com/36691
Reviewed-by: Bharath Vissapragada <bharathv@cloudera.com>
Tested-by: Bharath Vissapragada <bharathv@cloudera.com>