IMPALA-7530, IMPALA-7509. Re-plan queries if they fetch inconsistent metadata
Given the granular caching in CatalogdMetaProvider, it's possible to
have a situation like the following:

- impalad has cached the table list including some table "t"
- some other daemon has issued deletion of "t"
- we try to access "t" before the invalidation has reached us via the
  statestore

In this case, we'll get an error back when we try to fetch the table
"t". This can confuse the planning process since it previously
determined that the table exists. Note that this may occur either when
"t" is first referenced or later during planning (eg when fetching a
specific partition post-pruning), so it wouldn't be possible to simply
convert it into a 'table does not exist' result.

The solution here is to throw InconsistentMetadataFetchException after
invalidating the table list and associated object. This then triggers a
re-plan of the query.

This patch implements such re-planning up to 10 times before eventually
giving up. Given that each attempt invalidates the inconsistent cache we
would normally expect it to succeed on the first such retry. The limit
of 10 retries is just to avoid infinite loops in the case of a bug.

==C5_FEATURE_IMPALA_METADATA==

Change-Id: I17389954780fa22d7866224c17e128458fffa545
Reviewed-on: https://gerrit.sjc.cloudera.com/37228
Reviewed-by: Vuk Ercegovac <vercegovac@cloudera.com>
Tested-by: Vuk Ercegovac <vercegovac@cloudera.com>