KUDU-2395 (part 1): cache logged-in user
The name of the current logged-in user is not expected to change over
time once the process is running. If it did change this could cause
strange effects with ACL enforcement, for one, or result in outbound RPC
connections switching credentials during the process lifetime.

Additionally, as pointed at in KUDU-2395, the lookup of the user name
from the UID sometimes needs to go to external services which may be
slow and incur lock contention.

This patch switches to computing the local username only once and
reusing the result on all subsequent lookups.

Additional backport testing for 5.16: I added a DCHECK(false) in
GetLoggedInUser() and made sure that rpc-mgr-test hits it, confirming
that the method gets exercised by our tests.

Change-Id: Ib4500d89dd91e8eed6e4ae0661b886ab3c105130
Reviewed-on: http://gerrit.cloudera.org:8080/9895
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>
Reviewed-on: https://gerrit.sjc.cloudera.com/37231
Reviewed-by: Michael Ho <kwho@cloudera.com>
Tested-by: Lars Volker <lv@cloudera.com>