IMPALA-7591: Use short name for setting owner
This patch changes the owner to be the shortname in a Kerberized
cluster.  Authorization with Sentry is done using the shortname
and Hive uses the shortname.  Because object ownership privileges
are based on owner, authorization will not work with the full name.
The IOException for org.apache.impala.authorization.User.getShortName
should only ever be thrown if the format for the name is wrong.
i.e. It's actually a BadFormatString exception.

Testing:
- Ran all frontend, and e2e tests.
- Manually verified on kerberized cluster.

==C5_APPROVED_BUGFIX==

Change-Id: Iab392a0cd5c0bb98442e1f7ac39f470310a21f11
Reviewed-on: http://gerrit.cloudera.org:8080/11469
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit e34f28371e1942edb92d231e9588cda5c0f1cc50)
Reviewed-on: https://gerrit.sjc.cloudera.com/37356
Reviewed-by: Vuk Ercegovac <vercegovac@cloudera.com>
Tested-by: Vuk Ercegovac <vercegovac@cloudera.com>