IMPALA-7498: Fix log spew from LocalCatalog startup
Frontend calls LocalCatalog#waitForCatalog() in a tight loop during
startup and ends up spewing tons of log messages if the MetaProvider
takes some time to initialize.

This is a problem for CatalogdMetaProvider implementation since it
waits on the Catalog server to send an initial update. This can take
some time depending on various factors and the logs become too noisy
during that period.

This patch adds a sleep() inside waitForCatalog() to avoid this. We
can do something fancy, like ImpaladCatalog implementation, by waiting
on a synchronized monitor and getting notified when the Catalog is
initialized, but I don't think it is worth the effort, especially since
the sleep time is pretty small (MAX_CATALOG_UPDATE_WAIT_TIME_MS = 2s).

Testing: Tested it locally by just starting the impalad catalog in local
mode and the logs look much better while waiting for the Catalog to send
the initial update.

Change-Id: I81e7cc6f464fd71afecd94a5a2b777aff47ee11b
Reviewed-on: http://gerrit.cloudera.org:8080/11472
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Vuk Ercegovac <vercegovac@cloudera.com>