IMPALA-7606: Fix IllegalStateException in CatalogTableInvalidator
CatalogdTableInvalidator detects if a table is in a normal state using
Table.isLoaded() function. This is wrong because if there is an error
during the loading of a table, isLoaded() returns true. This patch
checks if the table is an IncompleteTable instead.
Also fixed a bug in tryInstallGcListener(). A test is added to test the
memory-based invalidation.

==C5_FEATURE_IMPALA_METADATA==

Change-Id: If938a40434b00af516445152f88832ef55d0d0ce
Reviewed-on: http://gerrit.cloudera.org:8080/11512
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>