IMPALA-7622: adds profile metrics when fetching incremental stats
Reapplies IMPALA-7622 after support for frontend profile metrics has
been reapplied.

When computing incremental statistics by fetching the stats directly
from catalogd, a potentially expensive RPC is made from the impalad
coordinator to catalogd. This change adds metrics to the frontend
section of the profile to track how long the request takes, the size
of the compressed bytes received, and the number of partitions received.

The profile for a 'compute incremental ...' command on a table with
no statistics looks like this:

Frontend:
     - StatsFetch.CompressedBytes: 0
     - StatsFetch.TotalPartitions: 24
     - StatsFetch.NumPartitionsWithStats: 0
     - StatsFetch.Time: 26ms

And the profile looks as follows when the table has stats, so the stats
are fetched:

Frontend:
     - StatsFetch.CompressedBytes: 24622
     - StatsFetch.TotalPartitions: 23
     - StatsFetch.NumPartitionsWithStats: 23
     - StatsFetch.Time: 14ms

Testing:
- manual inspection
- e2e test to check the profile

==C5_FEATURE_IMPALA_METADATA==

Change-Id: Ifce395e51044c0d00f1530863f1882152cbe9b90
Reviewed-on: https://gerrit.sjc.cloudera.com/38105
Reviewed-by: Vuk Ercegovac <vercegovac@cloudera.com>
Tested-by: Vuk Ercegovac <vercegovac@cloudera.com>
Reviewed-on: https://gerrit.sjc.cloudera.com/42642
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>