IMPALA-7581: timeout backend tests after 2 hours
This will abort the process if a backend test is taking too long, which
we assume is because of a hang. This makes job failures easier to triage
and may make it easier to debug failures if we collect coredumps or
minidumps.

Also disable the death tests for ASAN under the theory that the
probability of the hang is higher than a regular DEBUG build.

Testing:
Reduced the timeout to 5s and confirmed that it worked.

Change-Id: I2e4ef9cb0549ead0bae57b11489f6a4d9b44ef95
Reviewed-on: http://gerrit.cloudera.org:8080/11533
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>