IMPALA-7633: count_user_privilege isn't 0 at the end of test_owner
This patch adds a retry loop to validate the count of user privileges
in a SHOW GRANT USER statement after a DDL operation. The core of the
problem is cache consistency. When a DDL operation is executing, like
drop database, HMS is updated with the correct metadata, and Sentry is
updated to remove privileges from HMS. However, if a Sentry Refresh
happens between when HMS is updated CatalogOpExecutor:1322, and when
the local catalog privileges are updated CatalogOpExecutor:1341, then
the remove privilege call will fail and a log entry with "User does
not exist: foo_user" will be written to the log. The result is that
the response back to impalad with catalog updates will not contain
the user and privilege updates. Ultimately, when the "SHOW GRANT USER"
statement is run, it uses the local Impalad catalog which still
contains the privlege because it has not yet been updated from
statestore. This is not a security problem because the privilege
exists for a maximum of 2s by default, for an object that does not
exist. This is the same result as if the database was dropped from
Hive, except in that case it can be up to 62s by default that the
privilege exists for no object.

Testing:
- After retry was added, ran tests until log entry appeared and
  validate test did not fail.

==C5_APPROVED_BUGFIX==

Change-Id: Ifbba0fbd0e24a24b3f2af82ad5209f3fb7fb387b
Reviewed-on: http://gerrit.cloudera.org:8080/11595
Reviewed-by: Fredy Wijaya <fwijaya@cloudera.com>
Reviewed-by: Vuk Ercegovac <vercegovac@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit 2dda26420d36724793f096604033689fdfb05413)
Reviewed-on: https://gerrit.sjc.cloudera.com/37901
Tested-by: Fredy Wijaya <fwijaya@cloudera.com>
Reviewed-on: https://gerrit.sjc.cloudera.com/38126
Reviewed-on: https://gerrit.sjc.cloudera.com/42648
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>