IMPALA-5200: Count child time for parent's total time
One problem with the total time counter on runtime
profiles is that a parent's time may not be updated
if execution is stuck in a child node. The child
can accumulate time while the parent is stuck at
zero. This leads to incorrect or misleading
calculations of total time or non-child time
for the parent node during execution.

This makes a modest change in calculation for total
time for parent nodes. It takes advantage of the
fact that the parent should count all of the time
from all of its children as total time for itself.
Specifically, if a parent has accumulated X in its
total timer and its children have accumulated Y
summed across all of their timers, then a parent's
total time should be at least max(X, Y). There is no way
to know the appropriate overlap between X and Y,
so this uses a conservative calculation assuming
complete overlap.

This prevents a parent node from reporting itself
as 100% non-child time when it is actually stuck
executing child code. However, it does not help
if a child node is stuck and is not reporting its
own time.

Testing:
 - Added test case to runtime-profile-test
 - Core tests pass

Change-Id: Id6c1191c39fd18b6be45325366a74cf54908c77e
Reviewed-on: http://gerrit.cloudera.org:8080/11791
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>