IMPALA-7363: Fix buffer usage in ScannerContext::Stream::ReadVLong()
Previously, the ReadVLong() function would read a byte and use it
after subsequent calls to ReadBytes() function. As a result, the
value was invalidated depending on the contents of the
boundary_buffer_. This change ensures that all the functions on
the buffer are invoked before subsequent calls to ReadBytes().

Testing:
Enabled test_tpch_scan_ranges() for sequence files and ran it
multiple times to ensure it always produces the desired result.
Manually verified the buffer usage of other calls to ReadBytes()
and GetBytes().

Change-Id: Ic2b2ffe4b1d67c63cfcea2baedbeff48e65ca417
Reviewed-on: http://gerrit.cloudera.org:8080/11828
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>