IMPALA-7148: Make test_profile_fragment_instances() more robust
test_profile_fragment_instances() makes assumption about number
of instances of each scan node in the distributed query plan.
The number of instances for each scan nodes can actually vary
depending on how data is loaded and scheduler's decision.

This change relaxes the check for number of instances of each
scan node is a multiple of 3 which is the number of scan nodes
in the plan.

Change-Id: I08b068c21e9637575c85f4d54be9f7c56c106bf1
Reviewed-on: http://gerrit.cloudera.org:8080/11906
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>