IMPALA-7835: Role and user catalog objects with the same name can cause INVALIDATE METADATA to hang
Before this patch, enabling object ownership and running INVALIDATE
METADATA for the role and user catalog objects with the same principal
name could cause INVALIDATE METADATA to hang because a principal name
without a type is not guaranteed to be unique. This causes the catalog
delta logic to assume principal catalog objects with the same name but
with different types are considered the same causing a catalog version
inconsistency. This patch makes the principal object key to be unique by
appending the principal type information.

Testing:
- Added E2E authorization test
- Ran all FE tests
- Ran all E2E authorization tests

Change-Id: I516cf72e69e142a1349950cfca91f035c1ed445f
Reviewed-on: https://gerrit.sjc.cloudera.com/39443
Reviewed-by: Fredy Wijaya <fwijaya@cloudera.com>
Tested-by: Fredy Wijaya <fwijaya@cloudera.com>