IMPALA-6293: Unset JAVA_TOOL_OPTIONS variable from shell commands
JAVA_TOOL_OPTIONS is a variable used by JVM to specify the JVM options.
Impala has several flags that specify shell commands for Impala to run,
such as
- s3a_access_key_cmd
- s3a_secret_key_cmd
- ssl_private_key_password_cmd
- webserver_private_key_password_cmd

When debugging the JVM inside the Impala process, it is useful to
specify JAVA_TOOL_OPTIONS to run the Java debugger on a particular port.
However, JAVA_TOOL_OPTIONS remains in the environment, so it is passed
to these shell commands. If any of these shell commands run Java, then
that JVM will attempt to use the JAVA_TOOL_OPTIONS specified and thus
try to bind to the same port. The Impala process JVM is already bound to
that port, so this will fail.

This patch fixes the issue by unsetting JAVA_TOOL_OPTIONS before running
these shell commands.

Testing:
- Added a new test for os-util.cc

Change-Id: I96a84a27d3bf7b02f04e70562acfd67386b93183
Reviewed-on: http://gerrit.cloudera.org:8080/12005
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>