IMPALA-7473: fix crash when printing decimal with -v=3
The bug is that gcc emitted an aligned load instruction to
the int128 value_, and we didn't exercise that code path
in our testing. The fix is to use memcpy() to force an
unaligned load.

Also fix a NullPointerException in the frontend encountered
when running some tests with -v=3.

Testing:
Added a unit test that crashed when run with a release build.

Change-Id: I772681b78cda27d6c2086b91151ca8bdcd901828
Reviewed-on: http://gerrit.cloudera.org:8080/12029
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>