IMPALA-7903: Fix DCHECK failure in RawValue::PrintValue()
Occasionally Impala crashed on a DCHECK in RawValue::PrintValue()
when running web pages tests. Turned out that queries that have an
intermediate step for aggregation hit that DCHECK failure when the
log level is set to 3. As these web pages test run in parallel it
happened sometimes that a selective query was run after test_log_level
set the log level to 3 but before it came to the point to reset it.

The solution was to enhance RawValue::PrintValue() to handle
TYPE_FIXED_UDA_INTERMEDIATE not to run on the default case of the
switch.
In addition I extended the log writing for this DCHECK to print out
the type it can't handle.

Testing:
  - Extended the web server tests to set the log level to 3 and to
    run an aggregating query in the same test.

Change-Id: I22fe7cbbadd0c7eb6d0bd2d50bf500f0528e6af0
Reviewed-on: http://gerrit.cloudera.org:8080/12021
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>