Build parquet-reader earlier in test-with-docker
Test-with-docker builds the Impala container with the '-notests' switch
to save on build time and container size.
Since all EE_TEST suites/shards depend on the parquet-reader tool, all
EE_TEST containers start their life building this tool. This results
in all these containers hammering on CMake and the compiler, sometimes
in parallel.

The patch moves the line building parquet-reader into the "build"
phase of the Docker-based tests. The advantage is twofold:
- parquet-reader is built only once, saving some startup time for all
  the containers running tests
- the build also happens much faster at the end of the "build" phase,
  because the object files are still around and caches are hot
  (test-with-docker.py deletes all .o files before committing the container
  to shrink the container size that needs to be persisted).
  Building parquet-reader at the end of the build phase takes ~20 seconds,
  compared to the 1m20s it takes during the startup of a test container.

Verified by running test-with-docker.py on private infrastructure, and
checking build logs and test results -- still passing on Ubuntu 16.04

Change-Id: Iee141ad8b2a700378133a37498e74ddc306dfd57
Reviewed-on: http://gerrit.cloudera.org:8080/12025
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>