IMPALA-7928: Consistent remote read scheduling
Currently, remote reads for a particular file are not scheduled to
a consistent set of nodes. This reduces the efficiency of the HDFS
file handle cache.

This modifies the scheduling of remote reads to limit the number
of executors considered when picking an executor for a remote scan
range. The remote executor candidates are generated by hashing the
filename+offset multiple times and finding the closest nodes in a
hash ring. This is a consistent hash that is designed to limit the
number of files remapped when cluster nodes come and go. The number
of remote executor candidates is controlled by a query option
'num_remote_executor_candidates', which defaults to 3. It is
capped at 16.

Once the remote executor candidates are chosen, the algorithm for
picking a specific replica uses the same algorithm as picking a
local replica. It picks the node with the minimum number of
assigned bytes and uses 'schedule_random_replica' to determine
how to break ties.

This leaves the normal algorithms in place for local files, Kudu,
and HBase. If 'num_remote_executor_candidates' is set to 0, the
existing remote scheduling algorithm is used. The existing
algorithm schedules remote scan ranges on all available executors.

Testing:
 - There is a new hash-ring-test and related tests in scheduler-test.
 - There is a utility (hash-ring-util) in experiments for hand tuning
   the hash ring.

Change-Id: Icbf74088a8bd8c285ab7285ea3a01acd1bb53a45
Reviewed-on: http://gerrit.cloudera.org:8080/12037
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>