IMPALA-7183: Include remote host when logging unknown execution status.
ControlService::ReportExecStatus reports the execution status of
a fragment to the coordinator. When the coordinator receives a
ReportExecStatus call, it looks up the ClientRequestState for the query
id. If this cannot be found, usually because the query is in the process
of being canceled, then it logs a message before responding to the
RPC. Add the IP address and port of the remote host to this message.

Example output (from minicluster): "ReportExecStatus(): Received report
for unknown query ID (probably closed or cancelled):
174dbb16aa00a5c2:382e5cc800000000 remote host=127.0.0.1:47112"

TESTING
 Ran end-to-end tests.
 Forced code to execute by hand to check output is correct.

Change-Id: I79b223e22096e2d964b72654a34dc0d88e8c6422
Reviewed-on: http://gerrit.cloudera.org:8080/12102
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>