IMPALA-7265: Add parameter to cache remote HDFS file handles
Currently, the file handle cache does not apply to remote HDFS
files. This adds a parameter 'cache_remote_file_handles' that
enables the file handle cache for remote HDFS files. It is
currently being tested, so it is set to false by default.
This does not change the behavior for S3, ADLS, or ABFS.

Change-Id: I549f007432a01ca52fa8093d458a220bba02e1d9
Reviewed-on: http://gerrit.cloudera.org:8080/12111
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Philip Zeyliger <philip@cloudera.com>