IMPALA-8016: Fix lifecycle of classloader for UDFs.
The ClassLoader whence a UDF was loaded needs to be kept open for
executions of the UDF, so that the UDF can load other classes from the
same jar. (A typical scenario might be a utility class.) This was
broken by the fix to IMPALA-7668.

This commit moves closing the ClassLoader to the close() function.

A test for a UDF that imports a static method from another file has been
added. Doing so failed without this change.

Change-Id: Ic02e42fb25a2754ede21fe00312a60f07e0ba8a2
Reviewed-on: https://gerrit.sjc.cloudera.com/c/Impala/+/43718
Reviewed-by: Bharath Vissapragada <bharathv@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>