IMPALA-7446: enable buffer pool GC when near process mem limit
GC is performed when:
* The amount of memory allocated from the system for the buffer pool
  exceeds the reservation (i.e. free buffers and clean pages are not
  offset by unused reservation).
* The soft or hard process memory limit would otherwise cause an
  allocation to fail.

Testing:
Looped the old version of the semi_joins_exhaustive test, which
reliably reproduced the issue. I confirmed that the buffer pool GC was
running and that it preventing the query failures.

Added a backend test that reproed the issue. A large chunk of the code
change is to add infrastructure to use TCMalloc memory metrics
for the process memory tracker in backend tests.

Ran exhaustive tests.

Change-Id: I81e8e29f1ba319f1b499032f9518d32c511b4b21
Reviewed-on: http://gerrit.cloudera.org:8080/12133
Reviewed-by: Bikramjeet Vig <bikramjeet.vig@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>