IMPALA-8034: Improve planner tests
The FE PlannerTest cases are good, but often unrealistic and overly
complicated, especially when trying to verify selectivity and
cardinality. This commit adds new tests that isolate each bit of the
work for detailed inspection.

The current version of the tests highlighlights a number of bugs to be
fixed by ongoing work. This commit establishes a clear baseline of
current behavior, even if that behavior is not quite right. A "Bug:"
comment explains the expected result.

Tests: These are tests, no production code was changed.

Change-Id: I40e59e08d7ddf2b0391d42e50511aaf95d7275f4
Reviewed-on: http://gerrit.cloudera.org:8080/12145
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>