IMPALA-8043: Fix BE test failures related to SystemV timezones.
This is a fix for the following issue:

1. Some BE tests (e.g. ExprTest.TimestampFunctions) use the system's
   local timezone but run against a test timezone db (instead of the
   system's timezone db).
2. On some Linux installations /usr/share/zoneinfo contains symlinks
   to files in the /usr/share/zoneifo/SystemV directory
   (e.g /usr/share/zoneinfo/America/Los_Angeles is a symlink to
   ../SystemV/PST8PDT).
3. The 'SystemV' directory is not part of the test timezone db, since
   it is obsolete and excluded by default.

Consequently, if the system's local timezone is set to
America/Los_Angeles, BE tests won't find the corresponding timezone
file in the test timezone db. BE tests will default to UTC, which will
break some of them.

This change sets local timezone explicitly for failing BE tests, so
they don't depend on the system's local timezone.
It also adds 'SystemV' directory to the test timezone db to avoid
similar issues in the future.

Change-Id: I9288cd24c8af0c059e55d47c86bd92eaf0075681
Reviewed-on: http://gerrit.cloudera.org:8080/12199
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>