IMPALA-7929: Allow null qualifier in THBaseFilter
Impala query failed with "IntenalException: Required field 'qualifier'
was not present!" on table created via hive and mapped to HBase, because
the qualifier of the HBase key column is null in the mapped table, and
Impala required non-null qualifier. The fix here is to relax this
requirement.

Test:
Added unit test.
Tested in real cluster.

Change-Id: I378c2249604481067b5b1c3a3bbb28c30ad4d751
Reviewed-on: http://gerrit.cloudera.org:8080/12213
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Tim Armstrong <tarmstrong@cloudera.com>