IMPALA-8140: Fix use-after-poison in grouping aggregator
IMPALA-7333 changed the memory transfer semantics to move memory when
attaching it to the output batch instead of copying it. This caused a
use-after-poison when the cleaning up the hash tables during the call to
Close(). To fix this, we now clean up the hash table before closing the
output row stream.

Testing: added a test to aggregation.test

Change-Id: Id23cd1e2fc5e003e3c9e3503436621a76d49559d
Reviewed-on: http://gerrit.cloudera.org:8080/12298
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>